requirements.txt
flask==2.3.3
Pillow==10.0.0
requests==2.31.0  # Used for OpenAI API and GitHub API integration
openai==0.28.0
python-dotenv==1.0.0
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.0.76
jinja2==3.1.2